requirements.txt
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
enum34==1.1.10
Faker==4.1.2
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itchat==1.3.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.1
more-itertools==8.4.0
namedlist==1.7
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pytest==6.0.1
pytest-allure-adaptor==1.7.10
python-dateutil==2.8.1
requests==2.24.0
requests-toolbelt==0.9.1
scipy==1.5.2
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wxPython==4.1.0
zipp==3.1.0
